+++ runtime/actions.ts
@@ -32,9 +32,11 @@ import * as ACT_getPath    from "./datatypes/get-path";
 import * as ACT_object     from "./datatypes/object";
 import * as ACT_typeset    from "./datatypes/typeset";
 import * as ACT_pair       from "./datatypes/pair";
+import * as ACT_tuple      from "./datatypes/tuple";
 import * as ACT_map        from "./datatypes/map";
 import * as ACT_binary     from "./datatypes/binary";
 import * as ACT_series     from "./datatypes/series";
+import * as ACT_time       from "./datatypes/time";
 import * as ACT_tag        from "./datatypes/tag";
 import * as ACT_email      from "./datatypes/email";
 
@@ -326,11 +328,11 @@ module RedActions {
 		"ACT_HASH":       [ACT_block, ACT_series],
 		"ACT_PAIR":       [ACT_pair],
 		"ACT_PERCENT":    [ACT_float],
-		"ACT_TUPLE":      [],
+		"ACT_TUPLE":      [ACT_tuple],
 		"ACT_MAP":        [ACT_map],
 		"ACT_BINARY":     [ACT_binary, ACT_string, ACT_series],
 		"ACT_SERIES":     [ACT_series],
-		"ACT_TIME":       [],
+		"ACT_TIME":       [ACT_time],
 		"ACT_TAG":        [ACT_tag, ACT_string, ACT_series],
 		"ACT_EMAIL":      [ACT_email, ACT_string, ACT_series],
 		"ACT_HANDLE":     [],